Move get_transport to a transport module
This function is quite long and separating it to a module will help
clean up its relationship to config and the rest of the code.
